requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
matplotlib>=3.8.2
numpy>=1.26.3
pytest>=8.0.0
python-telegram-bot==20.7
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1 